pyproject.toml
[project]
name = "rk81"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.6",
    "aiolimiter>=1.2.1",
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "peewee>=3.18.1",
]
